+++ pkg/logql/engine.go
@@ -35,7 +35,8 @@ import (
 )
 
 const (
-	DefaultEngineTimeout = 5 * time.Minute
+	DefaultEngineTimeout       = 5 * time.Minute
+	DefaultBlockedQueryMessage = "blocked by policy"
 )
 
 var (
@@ -45,6 +46,13 @@ var (
 		Help:      "LogQL query timings",
 		Buckets:   prometheus.DefBuckets,
 	}, []string{"query_type"})
+
+	QueriesBlocked = promauto.NewCounterVec(prometheus.CounterOpts{
+		Namespace: "loki",
+		Name:      "blocked_queries",
+		Help:      "Count of queries blocked by per-tenant policy",
+	}, []string{"user"})
+
 	lastEntryMinTime = time.Unix(-100, 0)
 )
 
@@ -223,6 +231,7 @@ func (q *query) Exec(ctx context.Context) (logqlmodel.Result, error) {
 		if errors.Is(err, logqlmodel.ErrParse) ||
 			errors.Is(err, logqlmodel.ErrPipeline) ||
 			errors.Is(err, logqlmodel.ErrLimit) ||
+			errors.Is(err, logqlmodel.ErrBlocked) ||
 			errors.Is(err, context.Canceled) {
 			status = "400"
 		}
@@ -251,6 +260,10 @@ func (q *query) Eval(ctx context.Context) (promql_parser.Value, error) {
 		return nil, err
 	}
 
+	if q.checkBlocked(ctx, tenants) {
+		return nil, logqlmodel.ErrBlocked
+	}
+
 	switch e := expr.(type) {
 	case syntax.SampleExpr:
 		value, err := q.evalSample(ctx, e)
@@ -268,6 +281,19 @@ func (q *query) Eval(ctx context.Context) (promql_parser.Value, error) {
 	default:
 		return nil, errors.New("Unexpected type (%T): cannot evaluate")
 	}
+}
+
+func (q *query) checkBlocked(ctx context.Context, tenants []string) bool {
+	blocker := newQueryBlocker(ctx, q)
+
+	for _, tenant := range tenants {
+		if blocker.isBlocked(tenant) {
+			QueriesBlocked.WithLabelValues(tenant).Inc()
+			return true
+		}
+	}
+
+	return false
 }
 
 // evalSample evaluate a sampleExpr